refactor: extract InlineUtils to multiple files
